% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204187R.R
\name{hytest}
\alias{hytest}
\title{Testing Pareto versus Stretched-Exponential by wilks statistics}
\usage{
hytest(x, u, alpha)
}
\arguments{
\item{x}{the vector of samples}

\item{u}{the threshold of domain}

\item{alpha}{the significant level}
}
\value{
the result of the hypothesis \code{n}
}
\description{
Testing Pareto versus Stretched-Exponential by wilks statistics
}
\examples{
\dontrun{
  wtest <- hytest(c(2,3,4,5),1,0.05)
  wtest
}
}
