% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204187R.R
\name{Paretomle}
\alias{Paretomle}
\title{the mle of Pareto distribution}
\usage{
Paretomle(x, u)
}
\arguments{
\item{x}{the vector of samples}

\item{u}{the threshold of domain}
}
\value{
the mle of Pareto distribution \code{n}
}
\description{
the mle of Pareto distribution
}
\examples{
\dontrun{
  pmle <- Paretomle(c(2,3,4,5),1)
  pmle
}
}
