\name{SA23204187-package}
\alias{SA23204187-package}
\docType{package}
\title{
  Testing the Pareto Model versus the Stretched-Exponential Model
}
\description{
  Several R and Rcpp function are included in this package for hypothesis
  testing,including Wilks statistics et al.
}
\author{
Deyu He

Maintainer: Deyu He <he02@mail.ustc.edu.cn>
}
\keyword{ package }
