% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204187R.R
\name{SEmle}
\alias{SEmle}
\title{the approximate mle of Stretched-Exponential distribution}
\usage{
SEmle(x, u)
}
\arguments{
\item{x}{the vector of samples}

\item{u}{the threshold of domain}
}
\value{
the approximate mle of SE distribution \code{n}
}
\description{
the approximate mle of Stretched-Exponential distribution
}
\examples{
\dontrun{
  semle <- SEmle(c(2,3,4,5),1)
  semle
}
}
