% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wilks}
\alias{wilks}
\title{The wilks statistics of Pareto and Stretched-Exponential}
\usage{
wilks(x, u)
}
\arguments{
\item{x}{the vector of samples}

\item{u}{the threshold of domain}
}
\value{
the result of the hypothesis \code{n}
}
\description{
The wilks statistics of Pareto and Stretched-Exponential
}
\examples{
\dontrun{
wilkstat <- wilks(c(2,3,4,5),1)
wilkstat
}
}
